Re-org server structure
To make deploying the app easier (sticking everything into one
container) I want the backend to look for and serve assets in a similar
structure to what vite build will output. This way I can just copy over
the FE assets from the frontend container when building the production
image.

It's a bit janky, but workable for now.