Rename schema to avoid ESLint issues
There's no technical reason why we can't name a type and a value the
same thing (in fact, this is how zod's examples work) but ESLint doesn't
like it.

I'm ambivalent to the rule personally, but could see it being useful
outside of this specific type/value issue, so I'll leave it and rename
the zod schemas.