Accept &str instead of String
Because String can be coerced to &str using &str in the argument type
creates a more flexible api. Unfortunately, it ruins the simpler syntax
of just passing the method name to map, there might be a nicer way to
automatically reference the value, but I don't know about it at the
moment.