Add Satori requests
Unfortunately Satori doesn't have a public API, but using a session
token there is a "hidden" api that returns the data I want.

This means that they'll probably change the API without warning, but I
think this solution is still more reliable than scraping the HTML.

The cookies have a Max expiration set to 2038, so I shouldn't have to
update the cookie for quite a while (assuming that it doesn't get
destroyed if I log out).