Move caching to the individual request level
I want to be able to apply different TTLs to the redis keys based on the
data, so the first step is to cache the individual requests, instead of
the data sent to the frontend.

In theory this will make the APIs slower, but it should in the range of
milliseconds, so it shouldn't be noticeable.