Add proto definitions for the Anki API
Normally, you would dynamically generate the Rust types from the proto
files during build, but I didn't want to complicate the deploy process
for what (I'm hoping) is a small and static part of the app.

Therefore, I'm including the generated Rust code in the app, but I've
also kept the proto file in the event I need to re-generate.