Initial code to get Daily goal data
Satori includes a heat map on it's dashboard page, however the data is
included inside of a script tag and then loaded into the DOM after page
load.

I'd like to replace the regex with something better, ideally a JS
parser, but I don't know the state of that in Rust.