Switch to separate token and fetch auth header via login
I didn't realise when I added the frontend_api_token that it was a
session token, so it's going to expire quickly. To get around this I'm
now going through the login API to collect a new token. Unfortunately I
have to rely on the _grammar_app_session cookie, which is also session
based. If that also expires too quick then I'll have to try and figure
out a more permanent solution.