Change redis return value to be Option
Because the keys are set with a TTL after they expire the fetch request
will return nothing. However, since I defined the return type as a
String this was causing an error.

This commit doesn't actually change behaviour as caching is designed to
be optional so any cache errors are silently ignored, but it brings the
type signature inline with reality.